timm-mobilenet/requirements.txt
timm==1.0.25
torch
Pillow==12.1.1
matplotlib==3.10.8
numpy==2.4.3
